Move add_byte_to_accumulator to correct utils
