+++ src/cpu/opcodes/math/add.rs
@@ -1,4 +1,4 @@
-use cpu::utils::add_byte_to_accumulator;
+use cpu::opcodes::utils::add_byte_to_accumulator;
 use cpu::{Addressing, CPU};
 
 /// Add with carry
