+++ src/cpu/opcodes/mod.rs
@@ -6,3 +6,4 @@ pub mod registers;
 pub mod stack;
 pub mod storage;
 pub mod system;
+pub mod utils;
